trade/serializers.py
from rest_framework import serializers

from .models import Security, TradeHistory
from .services import import_data


class SecuritySerializer(serializers.ModelSerializer):
    class Meta:
        model = Security
        fields = "__all__"


class TradeHistorySerializer(serializers.ModelSerializer):
    class Meta:
        model = TradeHistory
        fields = "__all__"
