requirements.txt
django
djangorestframework
drf_yasg
gunicorn
requests
psycopg2-binary
dj_database_url